feat: move find to find
